refactor(NearComPromo): inline variant helper, restore ChainType colocation
Per review feedback from @jobotics: collapse the variant extraction back
into NearComPromo.tsx and restore ChainType inside useConnectWallet.ts.
The original extraction existed only to make the helper unit-testable;
with the test removed the extraction has no remaining purpose.

Also addresses CodeRabbit's perf nit by splitting NearComPromo into a
thin whitelabel-gated parent and a child that calls useConnectWallet(),
so non-near-intents templates no longer subscribe to wallet providers
for a banner they never render.
