--- src/hooks/chainType.ts
@@ -1,9 +1,0 @@
-export enum ChainType {
-  Near = "near",
-  EVM = "evm",
-  Solana = "solana",
-  WebAuthn = "webauthn",
-  Ton = "ton",
-  Stellar = "stellar",
-  Tron = "tron",
-}
